🛠[Refactor] 코드 리팩토링과 요구사항 변경에 따른 리팩토링

+++ feature/detail/src/main/java/com/android/myfooddiarybookaos/detail/mainUi/component/DetailTopLayer.kt
@@ -18,21 +18,25 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.android.myfooddiarybookaos.core.data.R
 import com.android.myfooddiarybookaos.data.state.AddScreenState
 import com.android.myfooddiarybookaos.data.utils.scaledSp
+import com.android.myfooddiarybookaos.detail.mainUi.popUp.DeleteDiaryPopUp
 import com.android.myfooddiarybookaos.detail.popup.DetailPopupScreen
 import com.android.myfooddiarybookaos.detail.viewModel.DetailViewModel
 import com.dnd_9th_3_android.gooding.data.root.ScreenRoot
 
 @Composable
 fun DetailTopLayer(
+    imageSize : Int,
     topDate: String,
     memoFixState: () -> Unit,
     detailViewModel: DetailViewModel = hiltViewModel()
 ) {
     val popUpState = remember { mutableStateOf(false) }
+    val deletePopupState = remember { mutableStateOf(false) }
 
     Box(
         Modifier
@@ -103,7 +107,7 @@ fun DetailTopLayer(
                                 ?.value = AddScreenState.FIX_IMAGE_IN_DETAIL
                             detailViewModel.appState.value
                                 ?.navController
-                                ?.navigate("${ScreenRoot.GALLERY}/false")
+                                ?.navigate("${ScreenRoot.GALLERY}/false/0")
                         },
                         addImage = {
                             popUpState.value = false
@@ -112,7 +116,7 @@ fun DetailTopLayer(
                                 ?.value = AddScreenState.ADD_IMAGE_IN_DETAIL
                             detailViewModel.appState.value
                                 ?.navController
-                                ?.navigate("${ScreenRoot.GALLERY}/true")
+                                ?.navigate("${ScreenRoot.GALLERY}/true/${imageSize}")
                         },
                         fixMemo = {
                             popUpState.value = false
@@ -120,9 +124,24 @@ fun DetailTopLayer(
                         },
                         deleteDiary = {
                             popUpState.value = false
-                            detailViewModel.deleteDiary()
+                            deletePopupState.value = true
                         }
                     )
+                }
+
+                if (deletePopupState.value){
+                    Dialog(onDismissRequest = {
+                        deletePopupState.value = false
+                    }) {
+                        DeleteDiaryPopUp(
+                            close = {
+                                deletePopupState.value = false
+                            },
+                            onDelete = {
+                                detailViewModel.deleteDiary()
+                            }
+                        )
+                    }
                 }
             }
         }
